🎨 chore : 디렉토리 변경, 코드 포맷팅
