+++ frontend/musta-project/src/components/order/OrderSellerItem.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { ListItem, ListItemText, Avatar, Typography, Card, Grid } from '@mui/material';
+import { ListItem, ListItemText, Avatar, Typography, Grid } from '@mui/material';
 import CheckCircleIcon from '@mui/icons-material/CheckCircle';
 import PendingIcon from '@mui/icons-material/Pending';
 import { styled } from '@mui/system';
 import { useNavigate } from 'react-router-dom';
 import ClearIcon from '@mui/icons-material/Clear';
 interface OrderItemProps {
-  order: OrderSellerWithArticle;
+  order: OrderSeller;
 }
 
 const StyledListItem = styled(ListItem)`
@@ -38,7 +38,7 @@ const UserInfoWrapper = styled(Grid)`
   align-items: flex-end;
 `;
 
-const OrderSellerWithArticleItem: React.FC<OrderItemProps> = ({ order }) => {
+const OrderSellerItem: React.FC<OrderItemProps> = ({ order }) => {
   const navigate = useNavigate();
 
   const handleItemClick = () => {
@@ -49,8 +49,7 @@ const OrderSellerWithArticleItem: React.FC<OrderItemProps> = ({ order }) => {
     <StyledListItem onClick={handleItemClick}>
       <UserAvatar alt={order.consumerName} src="/path/to/user-image.jpg" />
       <OrderItemText
-        primary={order.articleTitle}
-        secondary={order.consumerName}
+        primary={order.consumerName}
       />
       <UserInfoWrapper>
         <Typography variant="body2" color="textSecondary">
@@ -71,4 +70,4 @@ const OrderSellerWithArticleItem: React.FC<OrderItemProps> = ({ order }) => {
   );
 };
 
-export default OrderSellerWithArticleItem;
+export default OrderSellerItem;
