+++ frontend/musta-project/src/components/order/OrderSellerItemWithArticle.tsx
@@ -1,12 +1,12 @@
 import React from 'react';
-import { ListItem, ListItemText, Avatar, Typography, Card, Grid } from '@mui/material';
+import { ListItem, ListItemText, Avatar, Typography, Grid } from '@mui/material';
 import CheckCircleIcon from '@mui/icons-material/CheckCircle';
 import PendingIcon from '@mui/icons-material/Pending';
 import { styled } from '@mui/system';
 import { useNavigate } from 'react-router-dom';
 import ClearIcon from '@mui/icons-material/Clear';
 interface OrderItemProps {
-  order: OrderSeller;
+  order: OrderSellerWithArticle;
 }
 
 const StyledListItem = styled(ListItem)`
@@ -38,7 +38,7 @@ const UserInfoWrapper = styled(Grid)`
   align-items: flex-end;
 `;
 
-const OrderSellerItem: React.FC<OrderItemProps> = ({ order }) => {
+const OrderSellerWithArticleItem: React.FC<OrderItemProps> = ({ order }) => {
   const navigate = useNavigate();
 
   const handleItemClick = () => {
@@ -49,11 +49,12 @@ const OrderSellerItem: React.FC<OrderItemProps> = ({ order }) => {
     <StyledListItem onClick={handleItemClick}>
       <UserAvatar alt={order.consumerName} src="/path/to/user-image.jpg" />
       <OrderItemText
-        primary={order.consumerName}
+        primary={order.articleTitle}
+        secondary={order.consumerName}
       />
       <UserInfoWrapper>
         <Typography variant="body2" color="textSecondary">
-         {order.orderStatus}
+          {order.orderStatus}
         </Typography>
         <Typography variant="caption" color="textSecondary">
           {order.date}
@@ -70,4 +71,4 @@ const OrderSellerItem: React.FC<OrderItemProps> = ({ order }) => {
   );
 };
 
-export default OrderSellerItem;
+export default OrderSellerWithArticleItem;
