+++ frontend/musta-project/src/pages/orderPages/OrderSellerPage.tsx
@@ -1,7 +1,6 @@
-import OrderSellerItemList from '../components/OrderSellerItemList';
 import { Container, Paper, Typography } from '@mui/material';
 import { styled } from '@mui/system';
-import { makeObservable, observable, action } from 'mobx';
+import OrderSellerItemList from '../../components/order/OrderSellerItemList';
 
 const StyledContainer = styled(Container)`
   margin-top: 20px;
@@ -11,20 +10,6 @@ const StyledPaper = styled(Paper)`
   padding: 20px;
 `;
 
-
-class OrderSellerStore {
-  OrderSeller = null;
-  loading = false;
-
-  constructor() {
-    makeObservable(this, {
-      OrderSeller: observable,
-    });
-  }
-}
-
-const OrderStore = new OrderSellerStore();
-
 const OrderSellerPage = () => {
   return (
     <StyledContainer maxWidth="md">
@@ -32,7 +17,7 @@ const OrderSellerPage = () => {
         <Typography variant="h5" gutterBottom>
           나의 판매 목록
         </Typography>
-        <OrderSellerItemList  />
+        <OrderSellerItemList />
       </StyledPaper>
     </StyledContainer>
   );
