+++ frontend/musta-project/src/router/UserRouter.tsx
@@ -2,15 +2,15 @@
 import ArticlePage from '../pages/ArticlePage.tsx';
 import ReviewPage from '../pages/reviewPages/ReviewPage.tsx';
 import ReviewEditPage from '../pages/reviewPages/ReviewEditPage.tsx';
-import OrderDetailPage from "../pages/OrderDetailPage.tsx";
-import OrderConsumePage from "../pages/OrderConsumePage.tsx";
+import OrderConsumePage from "../pages/orderPages/OrderConsumePage.tsx";
 import { Routes, Route } from 'react-router-dom';
-import ArticleOrderPage from '../pages/AricleOrdersPage.tsx';
+import ArticleOrderPage from '../pages/orderPages/AricleOrdersPage.tsx';
 import Error404Page from '../pages/Error404Page.tsx';
 import HelloPage from '../pages/HelloPage.tsx';
 import { HomePage } from '../pages/HomePage.tsx';
 import MainPage from '../pages/MainPage.tsx';
-import OrderSellerPage from '../pages/OrderSellerPage.tsx';
+import OrderDetailPage from '../pages/orderPages/OrderDetailPage.tsx';
+import OrderSellerPage from '../pages/orderPages/OrderSellerPage.tsx';
 
 const UserRouter = () => {
   return (
